% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_mhcnuggets.R
\name{uninstall_mhcnuggets}
\alias{uninstall_mhcnuggets}
\title{Uninstall the MHCnuggets Python package.}
\usage{
uninstall_mhcnuggets(
  mhcnuggetsr_folder = mhcnuggetsr::get_default_mhcnuggets_folder(),
  ormr_folder_name = "python3"
)
}
\description{
Uninstall the MHCnuggets Python package.
}
\author{
Richèl J.C. Bilderbeek
}
